requirements.txt
arrow==1.2.3
better-profanity==0.7.0
blinker==1.5
cachelib==0.9.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.22.5
colorama==0.4.5
comtypes==1.1.11
cs50==9.2.2
docopt==0.6.2
face-recognition-models==0.3.0
Flask==2.2.2
Flask-Mail==0.9.1
Flask-QRcode==3.1.0
Flask-Session==0.4.0
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsons==1.6.3
MarkupSafe==2.1.1
numpy==1.23.3
Pillow==9.2.0
pipreqs==0.4.11
python-dateutil==2.8.2
python-dotenv==0.21.0
pyttsx3==2.90
qrcode==7.3.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
SpeechRecognition==3.8.1
SQLAlchemy==1.4.41
sqlparse==0.4.2
termcolor==1.1.0
threadpoolctl==3.1.0
typish==1.9.3
urllib3==1.26.12
Werkzeug==2.2.2
yarg==0.1.9
